feat(cli): redesign schedule commands to be agent-centric
This is a breaking change that simplifies the schedule workflow by making
agent name the primary identifier instead of schedule name.

Changes:
- Replace `vm0 schedule init` and `vm0 schedule deploy` with single
  `vm0 schedule setup <agent-name>` command that creates/edits schedules
- Update all commands (status, enable, disable, delete) to use
  `<agent-name>` as the argument instead of `<schedule-name>`
- Auto-generate schedule name as `{agent-name}-schedule` (hidden from user)
- Remove schedule.yaml file dependency - direct cloud deployment
- Add `--artifact-name` option to setup command (default: "artifact")
- Support both interactive wizard and non-interactive CLI options

BREAKING CHANGE: All schedule commands now require agent name instead of
schedule name. The schedule.yaml file is no longer used.

Closes #1627
