--- turbo/apps/cli/src/commands/schedule/deploy.ts
@@ -1,212 +1,0 @@
-import { Command } from "commander";
-import chalk from "chalk";
-import { existsSync, readFileSync } from "fs";
-import { parse as parseYaml } from "yaml";
-import { getComposeByName, deploySchedule } from "../../lib/api";
-import { scheduleYamlSchema, type ScheduleDefinition } from "@vm0/core";
-import { toISODateTime } from "../../lib/domain/schedule-utils";
-
-/**
- * Expand environment variables in a string
- * Supports ${VAR} syntax
- */
-function expandEnvVars(value: string): string {
-  return value.replace(/\$\{([^}]+)\}/g, (match, varName: string) => {
-    const envValue = process.env[varName];
-    if (envValue === undefined) {
-      console.warn(
-        chalk.yellow(`  Warning: Environment variable ${varName} not set`),
-      );
-      return match; // Keep original if not set
-    }
-    return envValue;
-  });
-}
-
-/**
- * Expand env vars in an object recursively
- */
-function expandEnvVarsInObject(
-  obj: Record<string, string> | undefined,
-): Record<string, string> | undefined {
-  if (!obj) return undefined;
-  const result: Record<string, string> = {};
-  for (const [key, value] of Object.entries(obj)) {
-    result[key] = expandEnvVars(value);
-  }
-  return result;
-}
-
-/**
- * Format an ISO date string in a specific timezone as YYYY-MM-DD HH:MM
- */
-function formatInTimezone(isoDate: string, timezone: string): string {
-  const date = new Date(isoDate);
-  const parts = new Intl.DateTimeFormat("en-CA", {
-    timeZone: timezone,
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-    hour12: false,
-  }).formatToParts(date);
-
-  const get = (type: string) => parts.find((p) => p.type === type)?.value ?? "";
-  return `${get("year")}-${get("month")}-${get("day")} ${get("hour")}:${get("minute")}`;
-}
-
-export const deployCommand = new Command()
-  .name("deploy")
-  .description("Deploy a schedule from schedule.yaml (create or update)")
-  .argument("[file]", "Path to schedule.yaml", "schedule.yaml")
-  .action(async (file: string) => {
-    try {
-      // Check if file exists
-      if (!existsSync(file)) {
-        console.error(chalk.red(`✗ File not found: ${file}`));
-        console.error(chalk.dim("  Create a schedule.yaml file first"));
-        process.exit(1);
-      }
-
-      // Read and parse YAML
-      const content = readFileSync(file, "utf-8");
-      let parsed: unknown;
-      try {
-        parsed = parseYaml(content);
-      } catch (err) {
-        console.error(chalk.red("✗ Invalid YAML syntax"));
-        if (err instanceof Error) {
-          console.error(chalk.dim(`  ${err.message}`));
-        }
-        process.exit(1);
-      }
-
-      // Validate schema
-      const result = scheduleYamlSchema.safeParse(parsed);
-      if (!result.success) {
-        console.error(chalk.red("✗ Invalid schedule.yaml format"));
-        for (const issue of result.error.issues) {
-          console.error(
-            chalk.dim(`  ${issue.path.join(".")}: ${issue.message}`),
-          );
-        }
-        process.exit(1);
-      }
-
-      const scheduleYaml = result.data;
-
-      // Process each schedule
-      const scheduleEntries = Object.entries(scheduleYaml.schedules) as [
-        string,
-        ScheduleDefinition,
-      ][];
-
-      if (scheduleEntries.length === 0) {
-        console.error(chalk.red("✗ No schedules defined in file"));
-        process.exit(1);
-      }
-
-      if (scheduleEntries.length > 1) {
-        console.error(
-          chalk.red("✗ Multiple schedules per file not supported yet"),
-        );
-        console.error(chalk.dim("  Please use one schedule per file"));
-        process.exit(1);
-      }
-
-      const [scheduleName, schedule] = scheduleEntries[0]!;
-
-      console.log(`Deploying schedule ${chalk.cyan(scheduleName)}...`);
-
-      // Resolve agent reference to compose ID
-      const agentRef = schedule.run.agent;
-
-      const namePart = agentRef.includes("/")
-        ? agentRef.split("/").pop()!
-        : agentRef;
-      const agentName = namePart.includes(":")
-        ? namePart.split(":")[0]!
-        : namePart;
-
-      const compose = await getComposeByName(agentName);
-      if (!compose) {
-        console.error(chalk.red(`✗ Agent not found: ${agentRef}`));
-        console.error(chalk.dim("  Make sure the agent is pushed first"));
-        process.exit(1);
-      }
-
-      const composeId = compose.id;
-
-      // Expand environment variables
-      const expandedVars = expandEnvVarsInObject(schedule.run.vars);
-      const expandedSecrets = expandEnvVarsInObject(schedule.run.secrets);
-
-      // Build deploy request
-      // Convert human-readable "YYYY-MM-DD HH:MM" format to ISO for the API
-      const atTime = schedule.on.at ? toISODateTime(schedule.on.at) : undefined;
-
-      const body = {
-        name: scheduleName,
-        composeId,
-        cronExpression: schedule.on.cron,
-        atTime,
-        timezone: schedule.on.timezone || "UTC",
-        prompt: schedule.run.prompt,
-        vars: expandedVars,
-        secrets: expandedSecrets,
-        artifactName: schedule.run.artifactName,
-        artifactVersion: schedule.run.artifactVersion,
-        volumeVersions: schedule.run.volumeVersions,
-      };
-
-      // Call API
-      const deployResult = await deploySchedule(body);
-
-      // Display result
-      if (deployResult.created) {
-        console.log(
-          chalk.green(`✓ Created schedule ${chalk.cyan(scheduleName)}`),
-        );
-      } else {
-        console.log(
-          chalk.green(`✓ Updated schedule ${chalk.cyan(scheduleName)}`),
-        );
-      }
-
-      // Show timezone
-      console.log(chalk.dim(`  Timezone: ${deployResult.schedule.timezone}`));
-
-      // Show trigger info based on type
-      if (deployResult.schedule.cronExpression) {
-        // Cron schedule: show cron expression and next run
-        console.log(
-          chalk.dim(`  Cron: ${deployResult.schedule.cronExpression}`),
-        );
-        if (deployResult.schedule.nextRunAt) {
-          const nextRun = formatInTimezone(
-            deployResult.schedule.nextRunAt,
-            deployResult.schedule.timezone,
-          );
-          console.log(chalk.dim(`  Next run: ${nextRun}`));
-        }
-      } else if (deployResult.schedule.atTime) {
-        // One-time schedule: show at time (no need for next run)
-        const atTime = formatInTimezone(
-          deployResult.schedule.atTime,
-          deployResult.schedule.timezone,
-        );
-        console.log(chalk.dim(`  At: ${atTime}`));
-      }
-    } catch (error) {
-      console.error(chalk.red("✗ Failed to deploy schedule"));
-      if (error instanceof Error) {
-        if (error.message.includes("Not authenticated")) {
-          console.error(chalk.dim("  Run: vm0 auth login"));
-        } else {
-          console.error(chalk.dim(`  ${error.message}`));
-        }
-      }
-      process.exit(1);
-    }
-  });
