+++ turbo/apps/cli/src/commands/schedule/index.ts
@@ -1,6 +1,5 @@
 import { Command } from "commander";
-import { initCommand } from "./init";
-import { deployCommand } from "./deploy";
+import { setupCommand } from "./setup";
 import { listCommand } from "./list";
 import { statusCommand } from "./status";
 import { deleteCommand } from "./delete";
@@ -10,8 +9,7 @@ import { disableCommand } from "./disable";
 export const scheduleCommand = new Command()
   .name("schedule")
   .description("Manage agent schedules")
-  .addCommand(initCommand)
-  .addCommand(deployCommand)
+  .addCommand(setupCommand)
   .addCommand(listCommand)
   .addCommand(statusCommand)
   .addCommand(deleteCommand)
